공과금 관련 위젯 구조 리팩터링
